refactor: remove Apache Kafka and Heroku Free
