+++ app.json
@@ -1,16 +1,12 @@
 {
     "name": "eCars Sample App",
     "description": "eCars - Salesforce Platform application demo",
-    "keywords": ["Salesforce", "Kafka", "Postgres", "MQTT", "LWC"],
+    "keywords": ["Salesforce", "Postgres", "MQTT", "LWC"],
     "website": "https://github.com/trailheadapps/ecars",
     "logo": "https://avatars0.githubusercontent.com/u/29558898?s=200&v=4",
     "env": {
         "APP_BASE": {
             "description": "Subdirectory within monorepo to deploy from"
-        },
-        "USE_KAFKA": {
-            "description": "Optionally include Kafka in the architecture. See the Demo Highlights section of the README for more details. https://github.com/trailheadapps/ecars#demo-highlights",
-            "value": "false"
         }
     },
     "buildpacks": [
