+++ apps/ecars-mqtt-broker/package.json
@@ -9,7 +9,7 @@
         "lint": "eslint '**/*.{js,ts}'"
     },
     "dependencies": {
-        "aedes": "^0.46.2",
+        "aedes": "^0.48.0",
         "mqtt": "^4.3.5",
         "websocket-stream": "^5.5.2"
     },
@@ -17,8 +17,8 @@
         "eslint": "^8.26.0"
     },
     "engines": {
-        "node": ">=10.13.0 <13.0.0",
-        "npm": ">=6.4.1 <7.0.0",
-        "yarn": ">=1.9.4 <2.0.0"
+        "node": ">=10.13.0 <19.0.0",
+        "npm": ">=6.4.1 <9.0.0",
+        "yarn": ">=1.9.4 <3.0.0"
     }
 }
