+++ apps/ecars-realtime/.env.sample
@@ -1,9 +1,1 @@
 DATABASE_URL=postgres://user:pass@host:port/databasename
-KAFKA_CLIENT_CERT=
-KAFKA_CLIENT_CERT_KEY=
-KAFKA_PREFIX=
-KAFKA_TRUSTED_CERT=
-KAFKA_URL=kafka+ssl://host1:9096,kafka+ssl://host2:9096,kafka+ssl://host3:9096,kafka+ssl://host4:9096,kafka+ssl://host5:9096,kafka+ssl://host6:9096,kafka+ssl://host7:9096,kafka+ssl://host8:9096
-KAFKA_TOPIC=sensor_data
-KAFKA_CONSUMER_GROUP=
-USE_KAFKA=
