+++ apps/ecars-realtime/Procfile
@@ -1,4 +1,3 @@
 web: node server.js
-sensor-persistence: node sensor-persistence.js
 sensor-simulator: node sensor-simulator.js
-sensor-connector: node sensor-connector.js
+sensor-persistence: node sensor-persistence.js
