+++ apps/ecars-realtime/README.md
@@ -36,56 +36,17 @@ node sensor-simulator.js
 
 ## MQTT Event Connector
 
-### Run locally
+### Configuration
+
+-   `DATABASE_URL` - Postgres connection string
+
+For local dev, create a file called `.env` at the top level directory with the above values defined. See `.env.sample` for the expected format. You can use Postgres running locally, or you can create a "development" Heroku app, add Postgres add-ons to it, and fill in `.env` with the environment variable values Heroku gives you. The latter is faster but requires internet access.
 
 ```
 npm install
 node sensor-connector.js
 # or use this instead to see lots of debug info
-# DEBUG='mqtt:*,connector,kafka-connect' node sensor-connector.js
-```
-
-## Kafka (Optional)
-
-If you are planning to use Kafka, you'll need to install the following paid addon:
-
-```
-heroku addons:create heroku-kafka:standard-0
-```
-
-Then, create a topic and a consumer-group for Kafka:
-
-```
-heroku kafka:topics:create <name> --partitions 1
-heroku kafka:consumer-groups:create <name>  # this is only necessary for `basic` kafka plans
-```
-
-### Configuration
-
--   `USE_KAFKA` - Enable Kafka
--   `KAFKA_CLIENT_CERT` - Contents of Kafka client cert
--   `KAFKA_CLIENT_CERT_KEY` - Contents of Kafka client key
--   `KAFKA_TRUSTED_CERT` - Contents of Kafka trusted cert (CA)
--   `KAFKA_PREFIX` - Prefix to prepend topic name and consumer group ID
--   `KAFKA_URL` - Comma-separated list of Kafka broker URLs
--   `KAFKA_TOPIC` - Kafka topic name
--   `KAFKA_CONSUMER_GROUP` - Kafka consumer group name
-
-## Events Persistence
-
-### Configuration
-
--   `DATABASE_URL` - Postgres connection string
-
-For local dev, create a file called `.env` at the top level directory with the above values defined. See `.env.sample` for the expected format. You can use Postgres and Kafka running locally, or you can create a "development" Heroku app, add Postgres and Kafka add-ons to it, and fill in `.env` with the environment variable values Heroku gives you. The latter is faster but requires internet access and Kafka is not free.
-
-### Run locally
-
-```
-npm install
-node sensor-persistence.js
-# or use this instead to see lots of debug info
-# DEBUG='kafka-consumer,kafka-connect:*,sequelize:*' node sensor-persistence.js
+# DEBUG='mqtt:*,connectort' node sensor-connector.js
 ```
 
 ## Troubleshooting
