+++ apps/ecars-realtime/packages/ecars-db/models/index.js
@@ -10,10 +10,7 @@ const db = {};
 
 let sequelize;
 if (config.use_env_variable) {
-    sequelize = new Sequelize(
-        process.env[config.use_env_variable] + '?ssl=true',
-        config
-    );
+    sequelize = new Sequelize(process.env[config.use_env_variable], config);
 } else {
     sequelize = new Sequelize(
         config.database,
