+++ apps/ecars-realtime/packages/ecars-events-connector/package.json
@@ -7,7 +7,6 @@
     "author": "Salesforce Developer Relations",
     "license": "CC0-1.0",
     "dependencies": {
-        "dotenv": "^16.0.3",
-        "sinek": "^9.1.0"
+        "dotenv": "^16.0.3"
     }
 }
