--- apps/ecars-realtime/packages/ecars-events-persistence/package.json
@@ -1,14 +1,0 @@
-{
-    "name": "@ecars/events-persistence",
-    "version": "1.0.0",
-    "private": true,
-    "description": "eCars - Kafka to Postgres service",
-    "main": "lib/kafka-consumer.js",
-    "author": "Salesforce Developer Relations",
-    "license": "CC0-1.0",
-    "dependencies": {
-        "debug": "^4.3.2",
-        "pino": "^7.6.5",
-        "sequelize-kafka-connect": "^4.0.0"
-    }
-}
