+++ apps/ecars-realtime/packages/ecars-mqtt-agent/package.json
@@ -7,8 +7,8 @@
     "author": "Salesforce Developer Relations",
     "license": "CC0-1.0",
     "dependencies": {
-        "debug": "^4.3.2",
-        "faker": "^5.5.3",
-        "mqtt": "^4.3.5"
+        "@faker-js/faker": "^7.6.0",
+        "debug": "^4.3.4",
+        "mqtt": "^4.3.7"
     }
 }
