--- apps/ecars-realtime/packages/ecars-utils/index.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const config = require('./config');
-const writeCerts = require('./write-certs');
-
-module.exports = {
-    config,
-    writeCerts
-};
