+++ apps/ecars-realtime/sensor-simulator.js
@@ -3,7 +3,7 @@
 require('dotenv').config();
 
 const Agent = require('@ecars/mqtt-agent');
-const faker = require('faker');
+const { faker } = require('@faker-js/faker');
 
 const SIMULATOR_CONCURRENCY = process.env.SIMULATOR_CONCURRENCY || 1;
 const SIMULATOR_INTERVAL = process.env.SIMULATOR_INTERVAL || 1500;
@@ -12,8 +12,8 @@ const profiles = ['short', 'medium', 'long'];
 
 async function main() {
     for (let i = 0; i < SIMULATOR_CONCURRENCY; i++) {
-        const agent = new Agent(profiles[faker.random.number(2)], {
-            name: `Pulsar One - ${faker.commerce.color()}`,
+        const agent = new Agent(profiles[faker.datatype.number(2)], {
+            name: `Pulsar One - ${faker.color.human()}`,
             interval: SIMULATOR_INTERVAL
         });
         try {
