+++ apps/ecars-services/package.json
@@ -20,31 +20,32 @@
         "dev": "tsc && npm run copy-files && concurrently -k -p \"[{name}]\" -n \"TypeScript,App\" -c \"yellow.bold,cyan.bold\"  \"tsc -w\" \"fastify start -w -l info -P dist/app.js\""
     },
     "dependencies": {
-        "@crcastle/fastify-piscina": "^2.0.0",
+        "@fastify/autoload": "^5.5.0",
+        "@fastify/helmet": "^10.0.2",
+        "ajv": "^8.11.2",
         "copyfiles": "^2.4.1",
-        "dotenv": "^16.0.0",
-        "fastify": "^3.27.1",
-        "fastify-autoload": "^3.10.0",
-        "fastify-cli": "^2.15.0",
-        "fastify-helmet": "^7.0.1",
-        "fastify-plugin": "^3.0.1",
+        "dotenv": "^16.0.3",
+        "fastify": "^4.10.2",
+        "fastify-cli": "^5.6.0",
+        "fastify-plugin": "^4.3.0",
         "jsforce": "^1.11.0",
-        "pdfmake": "^0.2.5",
-        "pg": "^8.7.3",
-        "web-push": "^3.4.5"
+        "pdfmake": "^0.2.6",
+        "pg": "^8.8.0",
+        "piscina": "^3.2.0",
+        "web-push": "^3.5.0"
     },
     "devDependencies": {
-        "@types/node": "^17.0.38",
-        "concurrently": "^7.3.0",
-        "json-schema-to-typescript": "^10.1.5",
-        "tap": "^15.1.6",
+        "@types/node": "^18.11.9",
+        "concurrently": "^7.6.0",
+        "json-schema-to-typescript": "^11.0.2",
+        "tap": "^16.3.2",
         "ts-node": "^10.9.1",
-        "typescript": "^4.7.4"
+        "typescript": "^4.9.3"
     },
     "engines": {
-        "node": ">=10.13.0 <13.0.0",
-        "npm": ">=6.4.1 <7.0.0",
-        "yarn": ">=1.9.4 <2.0.0"
+        "node": ">=10.13.0 <19.0.0",
+        "npm": ">=6.4.1 <9.0.0",
+        "yarn": ">=1.9.4 <3.0.0"
     },
     "tap": {
         "check-coverage": false
