+++ apps/ecars-services/tsconfig.json
@@ -3,8 +3,8 @@
         "module": "commonjs",
         "esModuleInterop": true,
         "allowSyntheticDefaultImports": true,
-        "target": "es2017",
-        "lib": ["es2017"],
+        "target": "es2020",
+        "lib": ["es2020"],
         "types": ["node"],
         "moduleResolution": "node",
         "sourceMap": true,
