+++ force-app/main/default/prompts/Case.prompt-meta.xml
@@ -3,7 +3,7 @@
     <masterLabel>Case</masterLabel>
     <promptVersions>
         <body
-        >Live car data is streamed from Kafka into Salesforce, providing real-time insights on the health of a car.</body>
+        >Live car data is streamed from MQTT into Salesforce, providing real-time insights on the health of a car.</body>
         <customApplication>ECars</customApplication>
         <delayDays>1</delayDays>
         <dismissButtonLabel>Close</dismissButtonLabel>
