+++ package.json
@@ -35,25 +35,25 @@
         "url": "git+https://github.com/trailheadapps/ecars"
     },
     "devDependencies": {
-        "@lwc/eslint-plugin-lwc": "^1.1.1",
+        "@lwc/eslint-plugin-lwc": "^1.5.0",
         "@prettier/plugin-xml": "^2.1.0",
-        "@sa11y/jest": "^3.1.0",
-        "@salesforce/eslint-config-lwc": "^3.2.1",
+        "@sa11y/jest": "^4.0.1",
+        "@salesforce/eslint-config-lwc": "^3.3.3",
         "@salesforce/eslint-plugin-lightning": "^1.0.0",
-        "@salesforce/sfdx-lwc-jest": "^1.1.0",
-        "@typescript-eslint/eslint-plugin": "^5.11.0",
-        "@typescript-eslint/parser": "^5.11.0",
-        "eslint": "^8.8.0",
-        "eslint-config-prettier": "^8.3.0",
+        "@salesforce/sfdx-lwc-jest": "^1.1.3",
+        "@typescript-eslint/eslint-plugin": "^5.44.0",
+        "@typescript-eslint/parser": "^5.44.0",
+        "eslint": "^8.28.0",
+        "eslint-config-prettier": "^8.5.0",
         "eslint-plugin-import": "^2.25.4",
-        "eslint-plugin-jest": "^26.1.0",
-        "husky": "^7.0.4",
-        "lerna": "^5.5.4",
-        "lint-staged": "^12.3.3",
-        "prettier": "^2.5.1",
-        "prettier-plugin-apex": "^1.10.0",
+        "eslint-plugin-jest": "^27.1.5",
+        "husky": "^8.0.2",
+        "lerna": "^6.0.3",
+        "lint-staged": "^13.0.3",
+        "prettier": "^2.7.1",
+        "prettier-plugin-apex": "^1.11.0",
         "rimraf": "^3.0.2",
-        "typescript": "^4.5.5"
+        "typescript": "^4.9.3"
     },
     "volta": {
         "node": "14.19.3",
